requirements.txt
mysql.connector
datetime
time
threading
twilio